Ex4.21

+++ controllers/blogs.js
@@ -40,6 +40,14 @@ blogsRouter.put('/:id', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+	const decodedToken = jwt.verify(request.token, config.SECRET)
+	if (!decodedToken.id) {
+		return response.status(401).json({ error: 'token missing or invalid' })
+	}
+	const blog = await Blog.findById(request.params.id)
+	if (blog.user.toString() !== decodedToken.id) {
+		return response.status(403).json({ error: 'forbidden - you can only delete your own blog' })
+	}
 	await Blog.findByIdAndRemove(request.params.id)
 	response.status(204).end()
 })
